+++ common/src/main/java/com/faboslav/friendsandfoes/common/block/CrabEggBlock.java
@@ -97,6 +97,7 @@ public final class CrabEggBlock extends Block
 					world.levelEvent(2001, pos, Block.getId(state));
 					CrabEntity crab = FriendsAndFoesEntityTypes.CRAB.get().create(world/*? >=1.21.3 {*/, VersionedEntitySpawnReason.BREEDING/*?}*/);
 					crab.setAge(-24000);
+					crab.setHome(crab.getNewHome());
 					crab.moveTo((double) pos.getX() + 0.3 + (double) j * 0.2, pos.getY(), (double) pos.getZ() + 0.3, 0.0F, 0.0F);
 					world.addFreshEntity(crab);
 				}
@@ -113,7 +114,6 @@ public final class CrabEggBlock extends Block
 		if (isSuitableBelow(world, pos) && !world.isClientSide()) {
 			world.levelEvent(2005, pos, 0);
 		}
-
 	}
 
 	private boolean shouldHatchProgress(Level world) {