+++ common/src/main/java/com/faboslav/friendsandfoes/common/mixin/LightningRodBlockMixin.java
@@ -212,7 +212,11 @@ public abstract class LightningRodBlockMixin extends LightningRodBlockBlockMixin
 	public void friendsandfoes_hasRandomTicks(
 		BlockState state, CallbackInfoReturnable<Boolean> cir
 	) {
-		cir.setReturnValue(true);
+		if(!FriendsAndFoes.getConfig().enableLightningRodOxidation) {
+			cir.setReturnValue(false);
+		} else {
+			cir.setReturnValue(true);
+		}
 	}
 
 	@Override